+++ RTL/pixel_memory_logo.v
@@ -1,7 +1,7 @@
 
 /*
 Author: Ali Ufuk Ercan
-Description: Reading, encoding and storing the pixel values from given file.
+Description: Reading and storing the pixel values from given file.
 Version: 1.00
 */
  
@@ -21,7 +21,7 @@ reg [1:0] memory [0:frame_height - 1] [0:frame_width - 1];
 
 
 initial begin
-    $readmemb("pgm_mikro_tasarim_string.txt",memory);
+    $readmemb("logo.txt",memory);
 end