+++ changelog.md
@@ -12,3 +12,11 @@
 
 # 1.02 - 2021-07-07
 - A proper combinational circuit is configured for the stretch register.
+
+</br>
+
+# 1.03 - 2021-07-12
+- Cubes pattern added.
+- 2 seperate block memory added for cubes and logo patterns.
+- Decoder and frame grabber modules added.
+- read.md updated.